+++ pkg/authenticator/config/validate_test.go
@@ -18,12 +18,12 @@ func TestValidate(t *testing.T) {
 		assert      errorAssertFunc
 	}{
 		{
-			description: "happy path",
+			description: "happy path - k8s",
 			settings: AuthnSettings{
 				// required variables
-				"CONJUR_AUTHN_URL":   "filepath",
+				"CONJUR_AUTHN_URL":   "authn-k8s",
 				"CONJUR_ACCOUNT":     "testAccount",
-				"CONJUR_AUTHN_LOGIN": "host",
+				"CONJUR_AUTHN_LOGIN": "host/myapp",
 				"MY_POD_NAME":        "testPodName",
 				"MY_POD_NAMESPACE":   "testNameSpace",
 				// correct value types
@@ -38,14 +38,48 @@ func TestValidate(t *testing.T) {
 			assert: assertEmptyErrorList(),
 		},
 		{
+			description: "happy path - jwt",
+			settings: AuthnSettings{
+				// required variables
+				"CONJUR_AUTHN_URL": "authn-jwt",
+				"CONJUR_ACCOUNT":   "testAccount",
+				"JWT_TOKEN_PATH":   "/tmp/token",
+				// correct value types
+				"CONJUR_CLIENT_CERT_RETRY_COUNT_LIMIT": "7",
+				"CONJUR_TOKEN_TIMEOUT":                 "6m0s",
+				"CONTAINER_MODE":                       "init",
+				// certificate provided
+				"CONJUR_SSL_CERTIFICATE": "samplecertificate",
+			},
+			assert: assertEmptyErrorList(),
+		},
+		{
+			description: "invalid jwt token path",
+			settings: AuthnSettings{
+				// required variables
+				"CONJUR_AUTHN_URL": "authn-jwt",
+				"CONJUR_ACCOUNT":   "testAccount",
+				// correct value types
+				"CONJUR_CLIENT_CERT_RETRY_COUNT_LIMIT": "7",
+				"CONJUR_TOKEN_TIMEOUT":                 "6m0s",
+				"CONTAINER_MODE":                       "init",
+				// certificate provided
+				"CONJUR_SSL_CERTIFICATE": "samplecertificate",
+				"JWT_TOKEN_PATH":         "invalid//path",
+			},
+			assert: assertErrorInList(fmt.Errorf(logger.CAKC065, "invalid//path")),
+		},
+		{
 			description: "error raised for missing required setting",
-			settings:    AuthnSettings{},
-			assert:      assertErrorInList(fmt.Errorf(logger.CAKC062, "CONJUR_AUTHN_URL")),
+			settings: AuthnSettings{
+				"CONJUR_AUTHN_URL": "authn-jwt",
+			},
+			assert: assertErrorInList(fmt.Errorf(logger.CAKC062, "CONJUR_ACCOUNT")),
 		},
 		{
 			description: "error raised for invalid username",
 			settings: AuthnSettings{
-				"CONJUR_AUTHN_URL":   "filepath",
+				"CONJUR_AUTHN_URL":   "authn-k8s",
 				"CONJUR_ACCOUNT":     "testAccount",
 				"CONJUR_AUTHN_LOGIN": "bad-username",
 				"MY_POD_NAME":        "testPodName",
@@ -56,7 +90,7 @@ func TestValidate(t *testing.T) {
 		{
 			description: "error raised for invalid retry count limit",
 			settings: AuthnSettings{
-				"CONJUR_AUTHN_URL":                     "filepath",
+				"CONJUR_AUTHN_URL":                     "authn-k8s",
 				"CONJUR_ACCOUNT":                       "testAccount",
 				"CONJUR_AUTHN_LOGIN":                   "host",
 				"MY_POD_NAME":                          "testPodName",
@@ -68,7 +102,7 @@ func TestValidate(t *testing.T) {
 		{
 			description: "error raised for invalid timeout",
 			settings: AuthnSettings{
-				"CONJUR_AUTHN_URL":                     "filepath",
+				"CONJUR_AUTHN_URL":                     "authn-k8s",
 				"CONJUR_ACCOUNT":                       "testAccount",
 				"CONJUR_AUTHN_LOGIN":                   "host",
 				"MY_POD_NAME":                          "testPodName",
@@ -81,7 +115,7 @@ func TestValidate(t *testing.T) {
 		{
 			description: "error raised for invalid certificate",
 			settings: AuthnSettings{
-				"CONJUR_AUTHN_URL":                     "filepath",
+				"CONJUR_AUTHN_URL":                     "authn-k8s",
 				"CONJUR_ACCOUNT":                       "testAccount",
 				"CONJUR_AUTHN_LOGIN":                   "host",
 				"MY_POD_NAME":                          "testPodName",
@@ -99,7 +133,8 @@ func TestValidate(t *testing.T) {
 	for _, tc := range TestCases {
 		t.Run(tc.description, func(t *testing.T) {
 			// SETUP & EXERCISE
-			errLogs := tc.settings.validate(&k8s.Config{}, successfulMockReadFile)
+			configObj, _ := getConfiguration(tc.settings["CONJUR_AUTHN_URL"])
+			errLogs := tc.settings.validate(configObj, successfulMockReadFile)
 
 			// ASSERT
 			tc.assert(t, errLogs)